+++ src/components/modals/next/index.js
@@ -1,8 +1,8 @@
 "use strict";
 
-var _interopRequireWildcard = require("@babel/runtime/helpers/interopRequireWildcard");
-
 var _interopRequireDefault = require("@babel/runtime/helpers/interopRequireDefault");
+
+var _interopRequireWildcard = require("@babel/runtime/helpers/interopRequireWildcard");
 
 Object.defineProperty(exports, "__esModule", {
   value: true
@@ -29,7 +29,7 @@ var _reactModal = _interopRequireDefault(require("react-modal"));
 
 var _closeButton = require("../../../components/buttons/close-button");
 
-var _button = require("../../../components/buttons/button/button");
+var _button = require("../../../components/buttons/button");
 
 var _next = require("../../../components/typography/paragraph/next");
 
@@ -49,15 +49,7 @@ var _gridConfig = require("../../../constants/grid-config");
 
 var _elementHelper = require("../../../helpers/dom/element-helper");
 
-function _templateObject() {
-  var data = (0, _taggedTemplateLiteral2.default)(["\n\n  body.k-ModalNext__body--open {\n    overflow: hidden;\n  }\n\n  .k-ModalNext__content {\n    position: relative;\n    background-color: ", ";\n    box-sizing: border-box;\n    transform: scale(0.94);\n    margin-right: ", ";\n    margin-left: ", ";\n    padding: ", " ", ";\n    width: calc(100vw ", ");\n    ", ";\n\n    @media (min-width: ", ") {\n      margin: auto;\n      padding: ", " ", ";\n      width: calc((100vw - ", ") + (", " * 11))\n    }\n\n    @media (min-width: ", ") {\n      padding: ", " ", ";\n      ", "\n    }\n\n    @media (min-width: ", ") {\n    ", "\n    }\n  }\n\n  .k-ModalNext__close {\n    position: absolute;\n    top: 0;\n    right: ", ";\n    @media (min-width: ", ") {\n      right: ", ";\n  }\n\n    button {\n      margin: 0;\n    }\n  }\n\n  .k-ModalNext__overlay {\n    position: fixed;\n    overflow: scroll;\n    display: flex;\n    flex-direction: column;\n    top: 0;\n    left: 0;\n    right: 0;\n    bottom: 0;\n    opacity: 0;\n    background-color: rgba(34, 34, 34, .8);\n    &::before ,\n    &::after {\n      content:'';\n      flex:1;\n      min-height: ", ";\n\n      @media (min-width: ", ") {\n        min-height: ", ";\n      }\n      ", "\n    }\n    ", "\n  }\n\n  .k-ModalNext__overlay--afterOpen {\n    transition: opacity .3s ease;\n    opacity: 1;\n  }\n  .k-ModalNext--afterOpen {\n    transition: opacity .3s ease, transform .3s ease;\n    transform: scale(1);\n    opacity: 1;\n  }\n\n  .k-ModalNext__overlay--beforeClose {\n    opacity: 0;\n  }\n  .k-ModalNext--beforeClose {\n    transition: opacity .3s ease, transform .5s ease;\n    transform: scale(1.06);\n    opacity: 0;\n  }\n\n  .k-ModalNext__title--fullSize {\n    position: sticky;\n    top:0;\n    width: 100vw;\n    text-align: center;\n    margin-left: ", ";\n    box-sizing: border-box;\n    background-color: ", ";\n    padding: ", " ", ";\n    border-bottom: ", " solid ", ";\n    margin-bottom: ", ";\n  }\n\n  .k-ModalNext__closeButton--fullSize {\n    position: absolute;\n    left: ", "\n    top: ", "\n  }\n"]);
-
-  _templateObject = function _templateObject() {
-    return data;
-  };
-
-  return data;
-}
+var _templateObject;
 
 var paddingPlusGutters = 2 * _gridConfig.CONTAINER_PADDING + 11 * _gridConfig.GUTTER;
 var oneGridCol = "calc((100vw - ".concat((0, _typography.pxToRem)(paddingPlusGutters), ") / 12 + ").concat((0, _typography.pxToRem)(_gridConfig.GUTTER), ")");
@@ -66,7 +58,7 @@ var StyledParagraph = (0, _styledComponents.default)(_next.Paragraph).withConfig
   displayName: "next__StyledParagraph",
   componentId: "sc-6re6vo-0"
 })(["font-size:", ";@media (min-width:", "){font-size:", ";}"], (0, _typography.pxToRem)(12), (0, _typography.pxToRem)(_screenConfig.ScreenConfig.S.min), (0, _typography.pxToRem)(14));
-var GlobalStyle = (0, _styledComponents.createGlobalStyle)(_templateObject(), _colorsConfig.default.background1, (0, _typography.pxToRem)(20), (0, _typography.pxToRem)(20), (0, _typography.pxToRem)(50), (0, _typography.pxToRem)(30), (0, _typography.pxToRem)(20), function (props) {
+var GlobalStyle = (0, _styledComponents.createGlobalStyle)(_templateObject || (_templateObject = (0, _taggedTemplateLiteral2.default)(["\n\n  body.k-ModalNext__body--open {\n    overflow: hidden;\n  }\n\n  .k-ModalNext__content {\n    position: relative;\n    background-color: ", ";\n    box-sizing: border-box;\n    transform: scale(0.94);\n    margin-right: ", ";\n    margin-left: ", ";\n    padding: ", " ", ";\n    width: calc(100vw ", ");\n    ", ";\n\n    @media (min-width: ", ") {\n      margin: auto;\n      padding: ", " ", ";\n      width: calc((100vw - ", ") + (", " * 11))\n    }\n\n    @media (min-width: ", ") {\n      padding: ", " ", ";\n      ", "\n    }\n\n    @media (min-width: ", ") {\n    ", "\n    }\n  }\n\n  .k-ModalNext__close {\n    position: absolute;\n    top: 0;\n    right: ", ";\n    @media (min-width: ", ") {\n      right: ", ";\n  }\n\n    button {\n      margin: 0;\n    }\n  }\n\n  .k-ModalNext__overlay {\n    position: fixed;\n    overflow: scroll;\n    display: flex;\n    flex-direction: column;\n    top: 0;\n    left: 0;\n    right: 0;\n    bottom: 0;\n    opacity: 0;\n    background-color: rgba(34, 34, 34, .8);\n    &::before ,\n    &::after {\n      content:'';\n      flex:1;\n      min-height: ", ";\n\n      @media (min-width: ", ") {\n        min-height: ", ";\n      }\n      ", "\n    }\n    ", "\n  }\n\n  .k-ModalNext__overlay--afterOpen {\n    transition: opacity .3s ease;\n    opacity: 1;\n  }\n  .k-ModalNext--afterOpen {\n    transition: opacity .3s ease, transform .3s ease;\n    transform: scale(1);\n    opacity: 1;\n  }\n\n  .k-ModalNext__overlay--beforeClose {\n    opacity: 0;\n  }\n  .k-ModalNext--beforeClose {\n    transition: opacity .3s ease, transform .5s ease;\n    transform: scale(1.06);\n    opacity: 0;\n  }\n\n  .k-ModalNext__title--fullSize {\n    position: sticky;\n    top:0;\n    width: 100vw;\n    text-align: center;\n    margin-left: ", ";\n    box-sizing: border-box;\n    background-color: ", ";\n    padding: ", " ", ";\n    border-bottom: ", " solid ", ";\n    margin-bottom: ", ";\n  }\n\n  .k-ModalNext__closeButton--fullSize {\n    position: absolute;\n    left: ", ";\n    top: ", ";\n  }\n"])), _colorsConfig.default.background1, (0, _typography.pxToRem)(20), (0, _typography.pxToRem)(20), (0, _typography.pxToRem)(50), (0, _typography.pxToRem)(30), (0, _typography.pxToRem)(20), function (props) {
   return props.fullSize && (0, _styledComponents.css)(["padding-top:0 !important;min-width:100vw !important;margin:0 !important;"]);
 }, (0, _typography.pxToRem)(_screenConfig.ScreenConfig.S.min), (0, _typography.pxToRem)(80), oneGridCol, (0, _typography.pxToRem)(paddingPlusGutters), (0, _typography.pxToRem)(_gridConfig.GUTTER), (0, _typography.pxToRem)(_screenConfig.ScreenConfig.L.min), (0, _typography.pxToRem)(80), oneGridCol, function (props) {
   return (0, _styledComponents.css)(["width:calc( ((100vw - ", ") / 12 + ", ") * ", " - ", " );"], (0, _typography.pxToRem)(paddingPlusGutters), (0, _typography.pxToRem)(_gridConfig.GUTTER), props.cols, (0, _typography.pxToRem)(_gridConfig.GUTTER));
@@ -194,20 +186,19 @@ var InnerModal = function InnerModal(_ref5) {
       modalProps = _ref5.modalProps,
       hasCloseButton = _ref5.hasCloseButton,
       maxWidth = _ref5.maxWidth,
-      big = _ref5.big,
-      huge = _ref5.huge,
+      size = _ref5.size,
       isOpen = _ref5.isOpen,
       zIndex = _ref5.zIndex,
       fullSize = _ref5.fullSize,
       fullSizeTitle = _ref5.fullSizeTitle,
-      others = (0, _objectWithoutProperties2.default)(_ref5, ["trigger", "children", "label", "labelledby", "describedby", "className", "closeButtonLabel", "onClose", "modalProps", "hasCloseButton", "maxWidth", "big", "huge", "isOpen", "zIndex", "fullSize", "fullSizeTitle"]);
+      others = (0, _objectWithoutProperties2.default)(_ref5, ["trigger", "children", "label", "labelledby", "describedby", "className", "closeButtonLabel", "onClose", "modalProps", "hasCloseButton", "maxWidth", "size", "isOpen", "zIndex", "fullSize", "fullSizeTitle"]);
 
   var _useContext3 = (0, _react.useContext)(ModalContext),
       _useContext4 = (0, _slicedToArray2.default)(_useContext3, 2),
       show = _useContext4[0].show,
       dispatch = _useContext4[1];
 
-  var colsOnDesktop = huge ? 10 : big ? 8 : 6;
+  var colsOnDesktop = size === 'huge' ? 10 : size === 'big' ? 8 : 6;
 
   var close = function close() {
     dispatch(updateState(false));
@@ -319,8 +310,7 @@ Modal.propTypes = {
   fullSize: _propTypes.default.bool,
   modalProps: _propTypes.default.object,
   hasCloseButton: _propTypes.default.bool,
-  big: _propTypes.default.bool,
-  huge: _propTypes.default.bool,
+  size: _propTypes.default.oneOf(['regular', 'big', 'huge']),
   isOpen: _propTypes.default.bool,
   zIndex: _propTypes.default.number,
   fullSizeTitle: _propTypes.default.string
@@ -333,8 +323,7 @@ Modal.defaultProps = {
   fullSize: false,
   modalProps: {},
   hasCloseButton: true,
-  big: false,
-  huge: false,
+  size: 'regular',
   isOpen: false,
   zIndex: 110,
   fullSizeTitle: ''