+++ src/components/uploaders/document-manager/index.js
@@ -25,7 +25,7 @@ var _colorsConfig = _interopRequireDefault(require("../../../constants/colors-co
 
 var _typography = require("../../../helpers/utils/typography");
 
-var _button = require("../../../components/buttons/button/button");
+var _button = require("../../../components/buttons/button");
 
 var _text = require("../../../components/typography/text");