+++ src/helpers/utils/lazy-hook.js
@@ -20,7 +20,7 @@ var useLazyObserver = function useLazyObserver(lazyComponentRef) {
       setLazyObserver = _useState2[1];
 
   (0, _react.useEffect)(function () {
-    var hasNoRef = !(lazyComponentRef === null || lazyComponentRef === void 0 ? void 0 : lazyComponentRef.current);
+    var hasNoRef = !(lazyComponentRef !== null && lazyComponentRef !== void 0 && lazyComponentRef.current);
 
     if (hasNoRef) {
       console.warn('lazyComponentRef.current does not exist, useLazyObserver will return true');